+++ thirdparty/tksheet/text_editor.py
@@ -2,18 +2,11 @@ from __future__ import annotations
 
 import tkinter as tk
 from collections.abc import Callable
-from typing import Literal
+from typing import Any, Literal
 
-from .functions import (
-    convert_align,
-)
-from .other_classes import (
-    DotDict,
-)
-from .vars import (
-    ctrl_key,
-    rc_binding,
-)
+from .constants import align_helper, ctrl_key
+from .functions import convert_align
+from .other_classes import DotDict, FontTuple
 
 
 class TextEditorTkText(tk.Text):
@@ -21,9 +14,9 @@ class TextEditorTkText(tk.Text):
         self,
         parent: tk.Misc,
         newline_binding: None | Callable = None,
+        rc_bindings: list[str] = "<3>",
     ) -> None:
-        tk.Text.__init__(
-            self,
+        super().__init__(
             parent,
             spacing1=0,
             spacing2=1,
@@ -37,7 +30,8 @@ class TextEditorTkText(tk.Text):
         self.newline_bindng = newline_binding
         self.rc_popup_menu = tk.Menu(self, tearoff=0)
         self.bind("<1>", lambda event: self.focus_set())
-        self.bind(rc_binding, self.rc)
+        for b in rc_bindings:
+            self.bind(b, self.rc)
         self.bind(f"<{ctrl_key}-a>", self.select_all)
         self.bind(f"<{ctrl_key}-A>", self.select_all)
         self.bind("<Delete>", self.delete_key)
@@ -53,6 +47,8 @@ class TextEditorTkText(tk.Text):
         font: tuple,
         bg: str,
         fg: str,
+        select_bg: str,
+        select_fg: str,
         state: str,
         text: str = "",
     ) -> None:
@@ -62,6 +58,8 @@ class TextEditorTkText(tk.Text):
             foreground=fg,
             insertbackground=fg,
             state=state,
+            selectbackground=select_bg,
+            selectforeground=select_fg,
         )
         self.editor_del_key = sheet_ops.editor_del_key
         self.align = align
@@ -69,45 +67,60 @@ class TextEditorTkText(tk.Text):
         self.rc_popup_menu.add_command(
             label=sheet_ops.select_all_label,
             accelerator=sheet_ops.select_all_accelerator,
+            image=sheet_ops.select_all_image,
+            compound=sheet_ops.select_all_compound,
             command=self.select_all,
             **menu_kwargs,
         )
         self.rc_popup_menu.add_command(
             label=sheet_ops.cut_label,
             accelerator=sheet_ops.cut_accelerator,
+            image=sheet_ops.cut_image,
+            compound=sheet_ops.cut_compound,
             command=self.cut,
             **menu_kwargs,
         )
         self.rc_popup_menu.add_command(
             label=sheet_ops.copy_label,
             accelerator=sheet_ops.copy_accelerator,
+            image=sheet_ops.copy_image,
+            compound=sheet_ops.copy_compound,
             command=self.copy,
             **menu_kwargs,
         )
         self.rc_popup_menu.add_command(
             label=sheet_ops.paste_label,
             accelerator=sheet_ops.paste_accelerator,
+            image=sheet_ops.paste_image,
+            compound=sheet_ops.paste_compound,
             command=self.paste,
             **menu_kwargs,
         )
         self.rc_popup_menu.add_command(
             label=sheet_ops.undo_label,
             accelerator=sheet_ops.undo_accelerator,
+            image=sheet_ops.undo_image,
+            compound=sheet_ops.undo_compound,
             command=self.undo,
             **menu_kwargs,
         )
-        align = convert_align(align)
-        if align == "w":
-            self.align = "left"
-        elif align == "e":
-            self.align = "right"
+        self.rc_popup_menu.add_command(
+            label=sheet_ops.redo_label,
+            accelerator=sheet_ops.redo_accelerator,
+            image=sheet_ops.redo_image,
+            compound=sheet_ops.redo_compound,
+            command=self.redo,
+            **menu_kwargs,
+        )
+        self.align = align_helper[convert_align(align)]
         self.delete(1.0, "end")
         self.insert(1.0, text)
         self.yview_moveto(1)
         self.tag_configure("align", justify=self.align)
         self.tag_add("align", 1.0, "end")
+        self.edit_reset()
 
-    def _proxy(self, command: object, *args) -> object:
+    def _proxy(self, command: Any, *args) -> Any:
         try:
             result = self.tk.call((self._orig, command) + args)
         except Exception:
@@ -119,16 +132,22 @@ class TextEditorTkText(tk.Text):
         ):
             self.tag_add("align", 1.0, "end")
             self.event_generate("<<TextModified>>")
-            if args and len(args) > 1 and args[1] != "\n" and args != ("1.0", "end"):
-                if self.yview() != (0.0, 1.0) and self.newline_bindng is not None:
-                    self.newline_bindng(check_lines=False)
+            if (
+                args
+                and len(args) > 1
+                and args[1] != "\n"
+                and args != ("1.0", "end")
+                and self.yview() != (0.0, 1.0)
+                and self.newline_bindng is not None
+            ):
+                self.newline_bindng(check_lines=False)
         return result
 
-    def rc(self, event: object) -> None:
+    def rc(self, event: Any) -> None:
         self.focus_set()
         self.rc_popup_menu.tk_popup(event.x_root, event.y_root)
 
-    def delete_key(self, event: object = None) -> None:
+    def delete_key(self, event: Any = None) -> None:
         if self.editor_del_key == "forward":
             return
         elif not self.editor_del_key:
@@ -141,26 +160,34 @@ class TextEditorTkText(tk.Text):
             self.delete("insert-1c")
             return "break"
 
-    def select_all(self, event: object = None) -> Literal["break"]:
-        self.tag_add(tk.SEL, "1.0", tk.END)
-        self.mark_set(tk.INSERT, tk.END)
+    def select_all(self, event: Any = None) -> Literal["break"]:
+        self.tag_add(tk.SEL, "1.0", "end-1c")
+        self.mark_set(tk.INSERT, "end-1c")
         # self.see(tk.INSERT)
         return "break"
 
-    def cut(self, event: object = None) -> Literal["break"]:
+    def cut(self, event: Any = None) -> Literal["break"]:
         self.event_generate(f"<{ctrl_key}-x>")
+        self.event_generate("<KeyRelease>")
         return "break"
 
-    def copy(self, event: object = None) -> Literal["break"]:
+    def copy(self, event: Any = None) -> Literal["break"]:
         self.event_generate(f"<{ctrl_key}-c>")
         return "break"
 
-    def paste(self, event: object = None) -> Literal["break"]:
+    def paste(self, event: Any = None) -> Literal["break"]:
         self.event_generate(f"<{ctrl_key}-v>")
+        self.event_generate("<KeyRelease>")
         return "break"
 
-    def undo(self, event: object = None) -> Literal["break"]:
+    def undo(self, event: Any = None) -> Literal["break"]:
         self.event_generate(f"<{ctrl_key}-z>")
+        self.event_generate("<KeyRelease>")
+        return "break"
+
+    def redo(self, event: Any = None) -> Literal["break"]:
+        self.event_generate(f"<{ctrl_key}-Shift-z>")
+        self.event_generate("<KeyRelease>")
         return "break"
 
 
@@ -169,9 +196,9 @@ class TextEditor(tk.Frame):
         self,
         parent: tk.Misc,
         newline_binding: None | Callable = None,
+        rc_bindings: list[str] = "<3>",
     ) -> None:
-        tk.Frame.__init__(
-            self,
+        super().__init__(
             parent,
             width=0,
             height=0,
@@ -183,7 +210,7 @@ class TextEditor(tk.Frame):
         self.parent = parent
         self.r = 0
         self.c = 0
-        self.tktext = TextEditorTkText(self, newline_binding=newline_binding)
+        self.tktext = TextEditorTkText(self, newline_binding=newline_binding, rc_bindings=rc_bindings)
         self.tktext.grid(row=0, column=0, sticky="nswe")
 
     def get(self) -> str:
@@ -207,8 +234,11 @@ class TextEditor(tk.Frame):
         show_border: bool,
         menu_kwargs: DotDict,
         sheet_ops: DotDict,
+        font: FontTuple,
         bg: str,
         fg: str,
+        select_bg: str,
+        select_fg: str,
         align: str,
         state: str,
         r: int = 0,
@@ -221,9 +251,11 @@ class TextEditor(tk.Frame):
             menu_kwargs=menu_kwargs,
             sheet_ops=sheet_ops,
             align=align,
-            font=menu_kwargs.font,
+            font=font,
             bg=bg,
             fg=fg,
+            select_bg=select_bg,
+            select_fg=select_fg,
             state=state,
             text=text,
         )